form일부구현
